+++ src/plugins/widget-react.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom'
 
 import Placeholder from '../components/placeholder'
 import AttributionEditor from '../components/attribution-editor'
+import BlockSchema from '../schema/block-meta'
 
 const Components = {
   placeholder: Placeholder,
@@ -13,7 +14,13 @@ export default class WidgetReact extends WidgetBase {
   static type () { return 'react' }
   constructor (options) {
     super(options)
-    this.onChange = onChange.bind(this)
+
+    const {type} = this.initialBlock
+    const schema = BlockSchema[type] || BlockSchema.default
+
+    this.onChange = (path, value) => {
+      onChange.call(this, path, value, schema)
+    }
 
     const props = {
       initialBlock: this.initialBlock,
@@ -21,13 +28,12 @@ export default class WidgetReact extends WidgetBase {
       imgfloConfig: this.ed.imgfloConfig
     }
 
-    const {type} = this.initialBlock
     let Component = Components[type] || Components.attribution
     this.mounted = ReactDOM.render(new Component(props), this.el)
   }
 }
 
-function onChange (path, value) {
+function onChange (path, value, schema) {
   // Mutate block
   let block = this.initialBlock
   let parent = this.initialBlock.metadata
@@ -35,6 +41,11 @@ function onChange (path, value) {
     parent = parent[path[i]]
   }
   parent[path[path.length - 1]] = value
+
+  // Regenerate html if needed
+  if (schema.makeHtml) {
+    block.html = schema.makeHtml(block.metadata, block.cover)
+  }
 
   // Send change up
   this.ed.updateMediaBlock(block)