+++ src/schema/block-meta.js
@@ -1,13 +1,18 @@
-// Used to define what fields are available in metadata editor
+/*
+ * Used to define what fields are available in metadata editor
+ */
 
-const blockMetaSchema = {
+import encode from '../util/encode'
+
+export default {
   image: {
     title: true,
     description: true,
     isBasedOnUrl: true,
     cover: true,
     author: true,
-    publisher: true
+    publisher: true,
+    makeHtml: makeFigure
   },
   video: {
     title: true,
@@ -15,15 +20,26 @@ const blockMetaSchema = {
     isBasedOnUrl: true,
     cover: true,
     author: true,
-    publisher: true
+    publisher: true,
+    makeHtml: makeFigure
   },
   quote: {
     title: false,
-    description: false,
+    description: true,
     isBasedOnUrl: true,
     cover: false,
     author: true,
-    publisher: true
+    publisher: true,
+    makeHtml: makeQuote
+  },
+  article: {
+    title: true,
+    description: true,
+    isBasedOnUrl: true,
+    cover: true,
+    author: true,
+    publisher: true,
+    makeHtml: makeArticle
   },
   default: {
     title: true,
@@ -35,4 +51,41 @@ const blockMetaSchema = {
   }
 }
 
-export default blockMetaSchema
+
+function makeFigure (metadata, cover) {
+  let htmlString = `<figure>`
+  if (cover && cover.src) {
+    htmlString += `<img src="${cover.src}">`
+  }
+  htmlString += makeTitleDescription('figcaption', metadata)
+  htmlString += `</figure>`
+  return htmlString
+}
+
+function makeQuote (metadata, _) {
+  return `<blockquote>${encode(metadata.description)}</blockquote>`
+}
+
+function makeArticle (metadata, cover) {
+  let htmlString = `<article>`
+  if (cover && cover.src) {
+    htmlString += `<img src="${cover.src}">`
+  }
+  htmlString += makeTitleDescription(null, metadata)
+  htmlString += `</article>`
+  return htmlString
+}
+
+function makeTitleDescription (tag, metadata) {
+  let htmlString = ``
+  if (metadata && metadata.title) {
+    htmlString += `<h1>${encode(metadata.title)}</h1>`
+  }
+  if (metadata && metadata.description) {
+    htmlString += `<p>${encode(metadata.description)}</p>`
+  }
+  if (tag && htmlString) {
+    htmlString = `<${tag}>${htmlString}</${tag}>`
+  }
+  return htmlString
+}